add new operation ProcessCanisterHttpInternal
